+++ FunctionsPrototypes.h
@@ -3,6 +3,6 @@
 long long int Sum(long long int Num1, long long int Num2);
 long long int Subtraction(long long int Num1, long long int Num2);
 long long int Multiplication(long long int Num1, long long int Num2);
-long long int Division(long long int Num1, long long int Num2);
+double Division(double Num1, double Num2);
 long long int Potentiation(long long int Num1, long long int Num2);
-long long int SquareRoot(long long int Num1);
+double SquareRoot(long long int Num1);
